requirements.txt
djangorestframework
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
Django==2.0.1
django-redis==4.10.0
importlib-metadata==0.20
kombu==4.6.4
more-itertools==7.2.0
Pillow==6.1.0
PyMySQL==0.9.3
pytz==2019.2
redis==3.3.8
sqlparse==0.3.0
vine==1.3.0
zipp==0.6.0
